+++ app/utils/functions.ts
